Require npm >= 3 and avoid the peerDependencies problem.
Use engineStrict to cause installation to fail with npm 2.x. Yes, I know engineStrict is deprecated, but that's as of npm 3.x and that's the version we need to require.

For now, we'll only verify with node 5.0 and stable. Would like to figure out how to have Travis verify node 4.0 + npm >= 3.0, but I don't see any such option at the moment.